Update README.md
Out of more than 300,000 projects launched on Kickstarter, only a third made it through the funding process with a positive outcome. Despite over $2 billion donation, not every project was successful. Getting funded on Kickstarter requires meeting or exceeding the project's initial goal, so many organizations spend months looking through past projects in an attempt to discover some trick for finding success. This is an analysis of a database of 4K past projects that is used to uncover any hidden trends or shortcomings.